Improved diagram fonts (#33)
The default Excalidraw font is too much like cursive. Changed to
schoolbell-v18-latin-regular and permanent-marker-v16-latin-regular for
headings/emphasis. I tested a lot of handwriting-style fonts and these
are the winners in terms of legibility and well suited to the Excalidraw
"rough" line style.

Minor improvements such as resizing a few things.